// src/models/RemoteAnnouncement.ts
export type RemoteAnnouncement = {
    rowNum: string;
    boardId: string;
    boardIdx: string;
    boardSubIdx: number;
    boardSubject: string;
    notice: string;
    memberType: string;
    memberIdx: string;
    memberEmail: string;
    memberName: string;
    hit: number;
    useTag: string;
    createdAt: string;
    boardContents: string;
    reply: [];
    new: string;
};